added option to process only PEs

+++ preprocess/preprocess_pipeline.bash
@@ -1,4 +1,3 @@
-
 ###############################################################################
 ### 1. Set env
 ###############################################################################
@@ -23,6 +22,8 @@ Usage: ./preprocess_pipeline.bash <options>
 --output_dir CHAR               directory to output generated data (i.e., preprocessed data, plots, tables)
 --overwrite t|f                 overwrite previous directory (default f)
 --pvalue NUM                    p value used to run pear. See pear help for valid p values (default: 0.01)
+--output_pe t|f                 output quality checked paired-end reads as fastq files (default f)
+--output_merged t|f             output quality checked merged reas as fasta files (i.e., workable.fasta) (default t)
 --reads CHAR                    input R1 reads
 --reads2 CHAR                   input R2 reads
 --sample_name CHAR              sample name (default metagenomex)
@@ -132,7 +133,33 @@ while :; do
   --overwrite=) # Handle the empty case
   printf 'Using default environment.\n' >&2
   ;;
- #############
+#############
+  --output_pe)
+  if [[ -n "${2}" ]]; then
+    OUTPUT_PE="${2}"
+    shift
+  fi
+  ;;
+  --output_pe=?*)
+  OUTPUT_PE="${1#*=}" # Delete everything up to "=" and assign the remainder.
+  ;;
+  --output_pe=) # Handle the empty case
+  printf 'Using default environment.\n' >&2
+  ;;  
+#############
+  --output_merged)
+  if [[ -n "${2}" ]]; then
+    OUTPUT_MERGED="${2}"
+    shift
+  fi
+  ;;
+  --output_merged=?*)
+  OUTPUT_MERGED="${1#*=}" # Delete everything up to "=" and assign the remainder.
+  ;;
+  --output_merged=) # Handle the empty case
+  printf 'Using default environment.\n' >&2
+  ;;    
+#############
   --pvalue)
   if [[ -n "${2}" ]]; then
     PVALUE="${2}"
@@ -266,6 +293,14 @@ if [[ -z "${OVERWRITE}" ]]; then
   OVERWRITE="f"
 fi
 
+if [[ -z "${OUTPUT_MERGED}" ]]; then
+  OUTPUT_MERGED="t"
+fi
+
+if [[ -z "${OUTPUT_PE}" ]]; then
+  OUTPUT_PE="f"
+fi
+
 if [[ -z "${PVALUE}" ]]; then
   PVALUE="0.01"
 fi  
@@ -273,7 +308,6 @@ fi
 if [[ -z "${SAMPLE_NAME}" ]]; then
   SAMPLE_NAME="metagenomex"
 fi  
-
 
 if [[ -z "${SUBSAMPLE}" ]]; then
   SUBSAMPLE="f"
@@ -471,147 +505,180 @@ if [[ "${TRIM_ADAPTERS}" == "t" ]]; then
 fi
 
 ###############################################################################
-### 10. Merge
+### 10. Quality trim paired-end reads
 ###############################################################################
 
-if [[ "${MERGER}" == "pear" ]]; then
+if [[ "${OUTPUT_PE}" == "t" ]]; then
 
-  "${pear}" \
-  -f "${R1}" \
-  -r "${R2}" \
-  -o "${OUTPUT_DIR}/${SAMPLE_NAME}" \
-  -j "${NSLOTS}" \
-  -p "${PVALUE}" \
-  --min-overlap "${MIN_OVERLAP}"
-
-  if [[ $? != 0 ]]; then
-    echo "merge with ${pear} failed"
-    exit 1
-  fi
-  
-  mv "${OUTPUT_DIR}/${SAMPLE_NAME}.assembled.fastq" \
-  "${OUTPUT_DIR}/${SAMPLE_NAME}.assembled-02.fastq"
-  mv "${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward.fastq" \
-  "${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward-02.fastq"
-  mv "${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse.fastq" \
-  "${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse-02.fastq"
-  mv "${OUTPUT_DIR}/${SAMPLE_NAME}.discarded.fastq" \
-  "${OUTPUT_DIR}/${SAMPLE_NAME}.discarded-02.fastq"
-  
-  R_ASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}.assembled-02.fastq"
-  R1_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward-02.fastq"
-  R2_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse-02.fastq"
-  R_DISCARD="${OUTPUT_DIR}/${SAMPLE_NAME}.discarded-02.fastq"
-
-  if [[ ! -s "${R_ASSEM}" ]]; then
-    echo "Failed merging: 0 merged reads with ${pear}"
-    exit 1
-  fi  
-  
-fi  
-
-if [[ "${MERGER}" == "bbmerge" ]]; then
-
-  R_ASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}.assembled-02.fastq"
-  R1_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward-02.fastq"
-  R2_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse-02.fastq"
-  INSERT_HIST="${OUTPUT_DIR}/${SAMPLE_NAME}_insert_hist.txt"
-  R_DISCARD="${OUTPUT_DIR}/${SAMPLE_NAME}.discarded-02.fastq"
-  
-  "${bbmerge}" \
-  in="${R1}" \
-  in2="${R2}" \
-  out="${R_ASSEM}" \
-  outu="${R1_UNASSEM}" \
-  outu2="${R2_UNASSEM}" \
-  ihist="${INSERT_HIST}"
-  
-  if [[ $? != 0 ]]; then
-    echo "merge with ${bbmerge} failed"
-    exit 1
-  fi
-  
-  touch "${R_DISCARD}"
-  
-fi
-
-###############################################################################
-### 11. Quality check: merged reads
-###############################################################################
-
-R_ASSEM_QC="${OUTPUT_DIR}/${SAMPLE_NAME}.assembled_qc-03.fastq"
-
-"${bbduk}" \
-in="${R_ASSEM}" \
-out="${R_ASSEM_QC}" \
-minlength=50 \
-threads="${NSLOTS}" \
-qtrim=rl \
-trimq="${MIN_QUAL}"
-
-if [[ $? -ne 0 ]]; then
-  echo "bbduk quality trimming assembled reads failed"
-  exit 1
-fi  
-
-###############################################################################
-### 12. Quality check: unmerged reads
-###############################################################################
-
-if [[ -s "${R1_UNASSEM}" ]]; then
-
-  R1_UNASSEM_QC="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward_qc-03.fastq"
-  R2_UNASSEM_QC="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse_qc-03.fastq"
+  echo "quality trimming PE reads ..."
+  R1_QC="${OUTPUT_DIR}/${SAMPLE_NAME}_R1_qc-02.fastq"
+  R2_QC="${OUTPUT_DIR}/${SAMPLE_NAME}_R2_qc-02.fastq"
 
   "${bbduk}" \
-  in="${R1_UNASSEM}" \
-  in2="${R2_UNASSEM}" \
-  out="${R1_UNASSEM_QC}" \
-  out2="${R2_UNASSEM_QC}" \
+  in="${R1}" \
+  in2="${R2}" \
+  out="${R1_QC}" \
+  out2="${R2_QC}" \
   minlength=50 \
   threads="${NSLOTS}" \
   qtrim=rl \
   trimq="${MIN_QUAL}"
 
   if [[ $? -ne 0 ]]; then
-    echo "bbduk quality trimming unassembled reads failed"
+    echo "bbduk quality trimming paired-end reads failed"
     exit 1
   fi  
-
 fi
 
 ###############################################################################
-### 13. Convert to fasta
+### 11. Merge
 ###############################################################################
 
-echo "converting to fasta ..."
+if [[ "${OUTPUT_MERGED}" == "t" ]]; then
 
-R_ASSEM_QC_FA="${OUTPUT_DIR}/${SAMPLE_NAME}.assembled_qc-03.fasta"
+  if [[ "${MERGER}" == "pear" ]]; then
 
-"${fq2fa}" "${R_ASSEM_QC}" > "${R_ASSEM_QC_FA}"
+    "${pear}" \
+    -f "${R1}" \
+    -r "${R2}" \
+    -o "${OUTPUT_DIR}/${SAMPLE_NAME}" \
+    -j "${NSLOTS}" \
+    -p "${PVALUE}" \
+    --min-overlap "${MIN_OVERLAP}"
 
-if [[ $? -ne 0 ]]; then
-  echo "fq2fa merged reads failed"
-  exit 1
-fi
-
-if [[ -s "${R1_UNASSEM_QC}" ]]; then
-
-  R1_UNASSEM_QC_FA="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward_qc-03.fasta"
-  R2_UNASSEM_QC_FA="${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse_qc-03.fasta"
+    if [[ $? != 0 ]]; then
+      echo "merge with ${pear} failed"
+      exit 1
+    fi
   
-  "${fq2fa}" "${R1_UNASSEM_QC}" > "${R1_UNASSEM_QC_FA}"
+    mv "${OUTPUT_DIR}/${SAMPLE_NAME}.assembled.fastq" \
+    "${OUTPUT_DIR}/${SAMPLE_NAME}_assembled-02.fastq"
+    
+    mv "${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.forward.fastq" \
+    "${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R1-02.fastq"
+    
+    mv "${OUTPUT_DIR}/${SAMPLE_NAME}.unassembled.reverse.fastq" \
+    "${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R2-02.fastq"
+    
+    mv "${OUTPUT_DIR}/${SAMPLE_NAME}.discarded.fastq" \
+    "${OUTPUT_DIR}/${SAMPLE_NAME}_discarded-02.fastq"
+  
+    R_ASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}_assembled-02.fastq"
+    R1_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R1-02.fastq"
+    R2_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R2-02.fastq"
+    R_DISCARD="${OUTPUT_DIR}/${SAMPLE_NAME}_discarded-02.fastq"
+
+    if [[ ! -s "${R_ASSEM}" ]]; then
+      echo "Failed merging: 0 merged reads with ${pear}"
+      exit 1
+    fi 
+    
+  fi
+ 
+  if [[ "${MERGER}" == "bbmerge" ]]; then
+
+    R_ASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}_assembled-02.fastq"
+    R1_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R1-02.fastq"
+    R2_UNASSEM="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R2-02.fastq"
+    INSERT_HIST="${OUTPUT_DIR}/${SAMPLE_NAME}_insert_hist.txt"
+    R_DISCARD="${OUTPUT_DIR}/${SAMPLE_NAME}_discarded-02.fastq"
+  
+    "${bbmerge}" \
+    in="${R1}" \
+    in2="${R2}" \
+    out="${R_ASSEM}" \
+    outu="${R1_UNASSEM}" \
+    outu2="${R2_UNASSEM}" \
+    ihist="${INSERT_HIST}"
+  
+    if [[ $? != 0 ]]; then
+      echo "merge with ${bbmerge} failed"
+      exit 1
+    fi
+  
+    touch "${R_DISCARD}"
+    
+  fi  
+
+  #############################################################################
+  ### 12. Quality check: merged reads
+  #############################################################################
+
+  R_ASSEM_QC="${OUTPUT_DIR}/${SAMPLE_NAME}_assembled_qc-03.fastq"
+
+  "${bbduk}" \
+  in="${R_ASSEM}" \
+  out="${R_ASSEM_QC}" \
+  minlength=50 \
+  threads="${NSLOTS}" \
+  qtrim=rl \
+  trimq="${MIN_QUAL}"
 
   if [[ $? -ne 0 ]]; then
-    echo "fq2fa unmerged forward reads failed"
+    echo "bbduk quality trimming assembled reads failed"
+    exit 1
+  fi  
+  
+  #############################################################################
+  ### 13. Quality check: unmerged reads
+  #############################################################################
+
+  if [[ -s "${R1_UNASSEM}" ]]; then
+
+    R1_UNASSEM_QC="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R1_qc-03.fastq"
+    R2_UNASSEM_QC="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R2_qc-03.fastq"
+
+    "${bbduk}" \
+    in="${R1_UNASSEM}" \
+    in2="${R2_UNASSEM}" \
+    out="${R1_UNASSEM_QC}" \
+    out2="${R2_UNASSEM_QC}" \
+    minlength=50 \
+    threads="${NSLOTS}" \
+    qtrim=rl \
+    trimq="${MIN_QUAL}"
+
+    if [[ $? -ne 0 ]]; then
+      echo "bbduk quality trimming unassembled reads failed"
+      exit 1
+    fi
+    
+  fi
+
+  #############################################################################
+  ### 14. Convert to fasta
+  #############################################################################
+
+  echo "converting to fasta ..."
+
+  R_ASSEM_QC_FA="${OUTPUT_DIR}/${SAMPLE_NAME}_assembled_qc-03.fasta"
+
+  "${fq2fa}" "${R_ASSEM_QC}" > "${R_ASSEM_QC_FA}"
+
+  if [[ $? -ne 0 ]]; then
+    echo "fq2fa merged reads failed"
     exit 1
   fi
-  
-  "${fq2fa}" "${R2_UNASSEM_QC}" > "${R2_UNASSEM_QC_FA}"
 
-  if [[ $? -ne 0 ]]; then
-    echo "fq2fa unmerged reverse reads failed"
-    exit 1
+  if [[ -s "${R1_UNASSEM_QC}" ]]; then
+ 
+    R1_UNASSEM_QC_FA="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R1_qc-03.fasta"
+    R2_UNASSEM_QC_FA="${OUTPUT_DIR}/${SAMPLE_NAME}_unassembled_R2_qc-03.fasta"
+  
+    "${fq2fa}" "${R1_UNASSEM_QC}" > "${R1_UNASSEM_QC_FA}"
+
+    if [[ $? -ne 0 ]]; then
+      echo "fq2fa unmerged forward reads failed"
+      exit 1
+    fi
+  
+    "${fq2fa}" "${R2_UNASSEM_QC}" > "${R2_UNASSEM_QC_FA}"
+
+    if [[ $? -ne 0 ]]; then
+      echo "fq2fa unmerged reverse reads failed"
+      exit 1
+    fi
+    
   fi
   
 fi
@@ -622,53 +689,37 @@ fi
 
 echo "computing statistics ..."
 
-if [[ -a "${R1_AT}" ]]; then
-
-  FILES=$(ls \
-  "${R1_INPUT}" "${R2_INPUT}" "${R1_AT}" "${R2_AT}" "${R_ASSEM}" "${R1_UNASSEM}" \
-  "${R2_UNASSEM}" "${R_DISCARD}" "${R_ASSEM_QC}" "${R1_UNASSEM_QC}" "${R2_UNASSEM_QC}")
-
-else
-
-  FILES=$(ls \
-  "${R1}" "${R2}" "${R_ASSEM}" "${R1_UNASSEM}" \
-  "${R2_UNASSEM}" "${R_DISCARD}" "${R_ASSEM_QC}" "${R1_UNASSEM_QC}" "${R2_UNASSEM_QC}")
-
-fi
+FILES=$(find "${OUTPUT_DIR}" -name "*.fastq")
 
 echo "${FILES}" | \
 while read FILE; do
 
-  if [[ -a "${FILE}" ]]; then
-  
-    N=$(count_fastq "${FILE}")
+   N=$(count_fastq "${FILE}")
   
     if [[ $? -ne 0 ]]; then
       echo "count_fastq failed for ${FILE}"
       exit 1
     fi
   
-    if [[ -s "${FILE}" ]]; then
+  if [[ -s "${FILE}" ]]; then
     
-      L=$(compute_mean_length "${FILE}")
+    L=$(compute_mean_length "${FILE}")
       
-      if [[ $? -ne 0 ]]; then
-        echo "compute_mean_length failed for ${FILE}"
-        exit 1
-      fi
-      
-    else 
-    
-      L=0
-      
+    if [[ $? -ne 0 ]]; then
+      echo "compute_mean_length failed for ${FILE}"
+      exit 1
     fi
+      
+  else 
     
-    echo -e "${SAMPLE_NAME}\t$(basename ${FILE})\tnum_seq\t${N}" >> \
-    "${OUTPUT_DIR}/stats_tmp.tsv"
-    echo -e "${SAMPLE_NAME}\t$(basename ${FILE})\tmean_length\t${L}" >> \
-    "${OUTPUT_DIR}/stats_tmp.tsv"
-  
-  fi 
+    L=0
+      
+  fi
+    
+  echo -e "${SAMPLE_NAME}\t$(basename ${FILE})\tnum_seq\t${N}" >> \
+  "${OUTPUT_DIR}/stats_tmp.tsv"
+  echo -e "${SAMPLE_NAME}\t$(basename ${FILE})\tmean_length\t${L}" >> \
+  "${OUTPUT_DIR}/stats_tmp.tsv"
   
 done
 
@@ -691,7 +742,7 @@ rm "${OUTPUT_DIR}/stats_tmp.tsv"
 echo "creating plot ..."
 
 Rscript --vanilla \
-"${PREPROCESS_DIR}/plots.R" \
+"${plots}" \
 "${OUTPUT_DIR}/stats.tsv" \
 "${OUTPUT_DIR}/stats_plots.png"
 
@@ -725,7 +776,6 @@ if [[ "${CLEAN}" == "t" ]]; then
       exit 1
     fi
   fi
-
   
   if [[ "${TRIM_ADAPTERS}" == "t" ]]; then
   
@@ -737,17 +787,30 @@ if [[ "${CLEAN}" == "t" ]]; then
     fi
   fi
   
-  rm "${R_ASSEM}" "${R1_UNASSEM}" "${R2_UNASSEM}" "${R_DISCARD}"
-  rm "${R_ASSEM_QC}" "${R1_UNASSEM_QC}" "${R2_UNASSEM_QC}"
-
-  if [[ $? -ne 0 ]]; then
-    echo "rm intermediate files failed"
-    exit 1
+  if [[ "${OUTPUT_MERGED}" == "t" ]]; then
+  
+    rm "${R_ASSEM}" "${R1_UNASSEM}" "${R2_UNASSEM}" "${R_DISCARD}"
+    rm "${R_ASSEM_QC}" "${R1_UNASSEM_QC}" "${R2_UNASSEM_QC}"
+    
+    if [[ $? -ne 0 ]]; then
+      echo "rm intermediate files failed"
+      exit 1
+    fi
+    
   fi
 fi
 
 ###############################################################################
-### 18. rename to workable
+### 18. Rename to workable
 ###############################################################################
 
-mv "${R_ASSEM_QC_FA}" "${OUTPUT_DIR}/${SAMPLE_NAME}_workable.fasta"
+if [[ "${OUTPUT_MERGED}" == "t" ]]; then
+  mv "${R_ASSEM_QC_FA}" "${OUTPUT_DIR}/${SAMPLE_NAME}_workable.fasta"
+fi
+
+###############################################################################
+### 19. End
+###############################################################################
+
+echo "preprocess_pipeline.bash exited with 0 errors"
+exit 0
